zomato_v2 + caching/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.4
redis==5.0.1
fastapi-cache2==0.2.1
